refactor(server): colocate team services by domain
