+++ packages/server/src/__tests__/agent-identity.test.ts
@@ -10,8 +10,8 @@ import {
   listAgents,
   suspendAgent,
 } from "../services/agents/identity.js";
-import { createMember } from "../services/member.js";
-import { createOrganization } from "../services/organization.js";
+import { createMember } from "../services/team/member.js";
+import { createOrganization } from "../services/team/organization.js";
 import { createAdminContext, createTestAdmin, useTestApp } from "./helpers.js";
 import { DEFAULT_ORG_ID } from "./setup.js";
