+++ packages/server/src/__tests__/agent-service-extra.test.ts
@@ -25,8 +25,8 @@ import {
   updateAgentSkills,
 } from "../services/agents/identity.js";
 import { ensureClientSupportsRuntimeProvider } from "../services/agents/runtime/binding.js";
-import { createMember } from "../services/member.js";
-import { createOrganization } from "../services/organization.js";
+import { createMember } from "../services/team/member.js";
+import { createOrganization } from "../services/team/organization.js";
 import { createAdminContext, useTestApp } from "./helpers.js";
 
 describe("agent service extra coverage", () => {
