+++ packages/server/src/__tests__/attachments-route.test.ts
@@ -17,7 +17,7 @@ import {
 import { MemoryAttachmentBlobStore } from "../services/attachment-blob-store.js";
 import { editMessage, lockFileAttachmentRefsIfPresent, sendMessage } from "../services/chat/message.js";
 import { validateMessageAttachmentRefs } from "../services/doc-snapshots.js";
-import { ensureMembership } from "../services/membership.js";
+import { ensureMembership } from "../services/team/membership.js";
 import { uuidv7 } from "../uuid.js";
 import { createAdminContext, createTestAdmin, useTestApp } from "./helpers.js";
