+++ packages/server/src/__tests__/gitlab-webhook-stage3.test.ts
@@ -22,8 +22,6 @@ import {
   putContextReviewerAssignment,
   putContextReviewerEnablement,
 } from "../services/context-tree/reviewer/settings.js";
-import { deleteMember } from "../services/member.js";
-import { deactivateMembership, MEMBER_STATUSES, reactivateMembership } from "../services/membership.js";
 import { putOrgSetting } from "../services/org-settings.js";
 import {
   createGitlabConnection,
@@ -41,6 +39,8 @@ import {
   suspendGitlabLinksForMembership,
 } from "../services/scm/gitlab/identities.js";
 import { applyGitlabPersonnelEvidence, normalizeGitlabWebhook } from "../services/scm/gitlab/webhook.js";
+import { deleteMember } from "../services/team/member.js";
+import { deactivateMembership, MEMBER_STATUSES, reactivateMembership } from "../services/team/membership.js";
 import { createTestAdmin, seedClient, seedHealthyAgentRuntime, useTestApp } from "./helpers.js";
 
 type App = ReturnType<ReturnType<typeof useTestApp>>;
