+++ packages/server/src/__tests__/me-multi-org.test.ts
@@ -5,7 +5,7 @@ import { invitationRedemptions } from "../db/schema/invitations.js";
 import { members } from "../db/schema/members.js";
 import { organizations } from "../db/schema/organizations.js";
 import { createAgent } from "../services/agents/identity.js";
-import { rotateInvitation } from "../services/invitation.js";
+import { rotateInvitation } from "../services/team/invitation.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestAdmin, seedClient, useTestApp } from "./helpers.js";
