+++ packages/server/src/__tests__/onboarding-readiness.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 import type { Database } from "../db/connection.js";
 import { agents } from "../db/schema/agents.js";
 import { listOrgsWithPersonalAgent, listOrgsWithUsableNonHumanAgent } from "../services/agents/access-control.js";
-import { ensureMembership } from "../services/membership.js";
+import { ensureMembership } from "../services/team/membership.js";
 import { createTestAdmin, useTestApp } from "./helpers.js";
 
 /**
