+++ packages/server/src/__tests__/org-context-activation.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 import { members } from "../db/schema/members.js";
 import { organizationSettings } from "../db/schema/organization-settings.js";
 import { putOrgSetting } from "../services/org-settings.js";
-import { createOrganization } from "../services/organization.js";
+import { createOrganization } from "../services/team/organization.js";
 import { createTestAdmin, useTestApp } from "./helpers.js";
 
 describe("org-scoped member Context activation", () => {
