+++ packages/server/src/__tests__/ws-admin-org-isolation.test.ts
@@ -5,7 +5,7 @@ import WebSocket from "ws";
 import { members } from "../db/schema/members.js";
 import { users } from "../db/schema/users.js";
 import { createAgent } from "../services/agents/identity.js";
-import { createOrganization, resolveDefaultOrgId } from "../services/organization.js";
+import { createOrganization, resolveDefaultOrgId } from "../services/team/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp } from "./helpers.js";
