+++ packages/server/src/__tests__/ws-admin-pulse-visibility.test.ts
@@ -6,7 +6,7 @@ import { members } from "../db/schema/members.js";
 import { users } from "../db/schema/users.js";
 import { broadcastToAdmins } from "../services/admin-broadcast.js";
 import { createAgent } from "../services/agents/identity.js";
-import { createOrganization } from "../services/organization.js";
+import { createOrganization } from "../services/team/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp, seedClient } from "./helpers.js";
