+++ packages/server/src/__tests__/ws-bind-runtime-mismatch.test.ts
@@ -9,7 +9,7 @@ import { clients } from "../db/schema/clients.js";
 import { members } from "../db/schema/members.js";
 import { users } from "../db/schema/users.js";
 import { createAgent } from "../services/agents/identity.js";
-import { resolveDefaultOrgId } from "../services/organization.js";
+import { resolveDefaultOrgId } from "../services/team/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp } from "./helpers.js";
