+++ packages/server/src/__tests__/ws-session-event.test.ts
@@ -14,7 +14,7 @@ import * as sessionEventService from "../services/chat/sessions/events.js";
 import * as contextTreeIoService from "../services/context-tree/io.js";
 import { buildLandingCampaignChatMetadata } from "../services/landing-campaigns/metadata.js";
 import { putOrgSetting } from "../services/org-settings.js";
-import { resolveDefaultOrgId } from "../services/organization.js";
+import { resolveDefaultOrgId } from "../services/team/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp } from "./helpers.js";
