+++ packages/server/src/api/orgs/identity.ts
@@ -1,7 +1,7 @@
 import { updateOrganizationSchema } from "@first-tree/shared";
 import type { FastifyInstance } from "fastify";
 import { requireOrgAdmin, requireOrgMembership } from "../../scope/require-org.js";
-import * as orgService from "../../services/organization.js";
+import * as orgService from "../../services/team/organization.js";
 
 /**
  * Class B — `/api/v1/orgs/:orgId` itself: read & rename the org row.
