+++ packages/server/src/services/oauth-bootstrap.ts
@@ -2,8 +2,8 @@ import { isKnownLandingCampaignSlug, parseAgentTemplateIntentPath } from "@first
 import { eq } from "drizzle-orm";
 import type { Database } from "../db/connection.js";
 import { users } from "../db/schema/users.js";
-import { findActiveByToken, recordRedemption } from "./invitation.js";
-import { createPersonalTeam, ensureMembership, pickPrimaryMembership } from "./membership.js";
+import { findActiveByToken, recordRedemption } from "./team/invitation.js";
+import { createPersonalTeam, ensureMembership, pickPrimaryMembership } from "./team/membership.js";
 
 export type ExternalAccountBootstrapUser = {
   userId: string;
