+++ packages/server/src/services/team/member.ts
@@ -2,20 +2,20 @@ import { randomBytes } from "node:crypto";
 import { AGENT_STATUSES, AGENT_TYPES, type CreateMember, type UpdateMember } from "@first-tree/shared";
 import bcrypt from "bcrypt";
 import { and, asc, desc, eq, inArray, max, ne } from "drizzle-orm";
-import type { Database } from "../db/connection.js";
-import { agents } from "../db/schema/agents.js";
-import { members } from "../db/schema/members.js";
-import { messages } from "../db/schema/messages.js";
-import { users } from "../db/schema/users.js";
-import { BadRequestError, ConflictError, NotFoundError } from "../errors.js";
-import { uuidv7 } from "../uuid.js";
-import { createAgent } from "./agents/identity.js";
-import { lockWatcherProjectionMemberMutation } from "./chat/membership/lock.js";
-import { lockWatcherProjectionForMemberChanges, recomputeWatcherChats } from "./chat/membership/watcher.js";
+import type { Database } from "../../db/connection.js";
+import { agents } from "../../db/schema/agents.js";
+import { members } from "../../db/schema/members.js";
+import { messages } from "../../db/schema/messages.js";
+import { users } from "../../db/schema/users.js";
+import { BadRequestError, ConflictError, NotFoundError } from "../../errors.js";
+import { uuidv7 } from "../../uuid.js";
+import { createAgent } from "../agents/identity.js";
+import { lockWatcherProjectionMemberMutation } from "../chat/membership/lock.js";
+import { lockWatcherProjectionForMemberChanges, recomputeWatcherChats } from "../chat/membership/watcher.js";
+import { forceDisconnect } from "../runtime/connection-manager.js";
+import * as presenceService from "../runtime/presence.js";
+import { suspendGitlabLinksForMembership } from "../scm/gitlab/identities.js";
 import { MEMBER_STATUSES, reactivateMembership, syncUserDisplayName } from "./membership.js";
-import { forceDisconnect } from "./runtime/connection-manager.js";
-import * as presenceService from "./runtime/presence.js";
-import { suspendGitlabLinksForMembership } from "./scm/gitlab/identities.js";
 
 const SALT_ROUNDS = 10;
